+++ tests/_stubs/ExtendingUnknownClass.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Tests\Stub;
+
+use HiddenPlace\UnknownClass;
+
+class ExtendingUnknownClass extends UnknownClass
+{
+}